+++ spec/Handler/RepositoryBasedTransitionHandlerSpec.php
@@ -241,21 +241,17 @@ class RepositoryBasedTransitionHandlerSpec extends ObjectBehavior
 
     function it_transits_to_next_state(Transition $transition, Item $item, State $state)
     {
-        $item->transit($transition, Argument::type(Context::class), true)
-            ->shouldBeCalled()
-            ->willReturn($state);
+        $item->releaseRecordedStateChanges()
+            ->shouldBeCalledOnce()
+            ->willReturn([$state]);
 
         $transition->validate($item, Argument::type(Context::class))
             ->willReturn(true)
             ->shouldBeCalled();
 
-        $transition->executeActions($item, Argument::type(Context::class))
-            ->willReturn(true)
-            ->shouldBeCalled();
-
-        $transition->executePostActions($item, Argument::type(Context::class))
-            ->willReturn(true)
-            ->shouldBeCalled();
+        $transition->execute($item, Argument::type(Context::class))
+            ->willReturn($state)
+            ->shouldBeCalledOnce();
 
         $transition->checkCondition($item, Argument::type(Context::class))
             ->willReturn(true)
@@ -266,7 +262,7 @@ class RepositoryBasedTransitionHandlerSpec extends ObjectBehavior
             ->shouldBeCalled();
 
         $this->validate([]);
-        $this->transit()->shouldHaveType('Netzmacht\Workflow\Flow\State');
+        $this->transit()->shouldHaveType(State::class);
     }
 
     function it_checks_if_transition_is_available(Transition $transition, Item $item)